django app/requirements.txt
Django==3.0.2; python_version > '3.5'
Django==2.2.9; python_version == '3.5'
Django==1.11.27; python_version < '3.0'
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-cloud-datastore==2.0.0
